Fix layout of device name
