+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui_VolumeScreenIndividualTest_volumeScreenAtMinimum.png
