+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui_VolumeScreenThemeTest_volumeScreenThemes[1]_blue7fcfff.png
