+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui_VolumeScreenThemeTest_volumeScreenThemes[3]_green6dd58c.png
