+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui_VolumeScreenThemeTest_volumeScreenThemes[5]_orangey.png
