+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui_VolumeScreenThemeTest_volumeScreenThemes[6]_uamp.png
